unmarshal into temp map to keep deleted keys deleted
Unmarshal's behaviour is to keep the existing keys of a map, not to
reallocate a new map and fill with the new values being unmarshaled!

The docs say the following about encoding/json.Unmarshal:

»To unmarshal a JSON object into a map, Unmarshal first establishes a
map to use. If the map is nil, Unmarshal allocates a new map. Otherwise
Unmarshal reuses the existing map, ***keeping existing entries***." <-- uh oh